logout user(sistema e plataforma)

+++ cinematic-front/src/app/components/sistema/header-sistema/header-sistema.component.ts
@@ -2,6 +2,7 @@ import { Component, inject, Input, OnInit } from '@angular/core';
 import { IUsuario } from '../../../model/IUsuario';
 import { UsuarioService } from '../../../services/usuario/usuario.service';
 import { Router } from '@angular/router';
+import { ITipoUsuario } from '../../../model/ITipoUsuario';
 
 @Component({
   selector: 'app-header-sistema',
@@ -34,7 +35,16 @@ export class HeaderSistemaComponent implements OnInit{
   }
 
   logout(): void {
-    sessionStorage.clear();
-    this.router.navigate(['sistema/login']);
+    let tipoUser: string | null = sessionStorage.getItem('usuarioTipo');
+    console.log(tipoUser);
+    if (tipoUser!=null) {
+      if (tipoUser==='FUNCIONARIO' || tipoUser==='GERENTE') {
+        this.router.navigate(['sistema/login']);
+      }
+      if (tipoUser==='CLIENTE') {
+        this.router.navigate(['login']);
+      }
+      sessionStorage.clear();
+    }
   }
 }
